% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covcomp.R
\name{Compln}
\alias{Compln}
\title{Completion}
\usage{
Compln(Cov, Omega, VectorOfTuning)
}
\arguments{
\item{Cov}{The partial covariance matrix with same dimensions as Omega.}

\item{Omega}{A matrix of 0s and 1s; indicating the domain of partial covariance.}

\item{VectorOfTuning}{vector of truncation parameters with dimension at least m-1 where Omega is an m-serrated domain}
}
\value{
A matrix of the same dimensions: the covariance completion of Cov
}
\description{
Calculates the canonical completion of partial covariance assuming the domain is given by Omega.
}
