% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covcomp.R
\name{IndexCof}
\alias{IndexCof}
\title{C Index}
\usage{
IndexCof(Omega)
}
\arguments{
\item{Omega}{A matrix of 0s and 1s.}
}
\value{
Vector; range of C Index
}
\description{
Finds the C Index of Omega.
}
\examples{
Omega <- matrix(c(1,1,0,1,1,1,0,1,1), ncol = 3)
IndexCof(Omega)
}
