% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covcomp.R
\name{Pivot}
\alias{Pivot}
\title{Pivot}
\usage{
Pivot(Omega)
}
\arguments{
\item{Omega}{A matrix of 0s and 1s.}
}
\value{
Vector of dimension 2: entries of the pivot.
}
\description{
Finds the first pivot of Omega.
}
\examples{
Omega <- matrix(c(1,1,0,1,1,1,0,1,1), ncol = 3)
Pivot(Omega)
}
